+++ views/login.ejs
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+    <title>Document</title>
+</head>
+<body>
+    <button id='sign_in_visible_btn' onclick="signInVisible()"> 로그인 </button>
+    <button id='sign_up_visible_btn' onclick="signUpVisible()"> 회원가입 </button>
+    <div id="sign_in_div">
+        <div>로그인</div>
+        <form name="signin-form">
+            <input type="text" name="userid" placeholder="아이디"><br/>
+            <input type="password" name="pw" placeholder="비밀번호"><br/>
+            <button type="button" onclick="userSignin()">로그인</button><br/>
+        </form>
+        <a href="/findpw">비밀번호 찾기</a>
+    </div>
+    <div id="sign_up_div" hidden>
+        <div>회원가입</div>
+        <form name="signup-form">
+            <input type="text" name="name" placeholder="이름"/><br/>
+            <input type="text" name="userid" placeholder="아이디"><br/>
+            <input type="password" name="pw" placeholder="비밀번호"><br/>
+            <input type="email" name="email" placeholder="임시 비밀번호를 받을 email"><br/>
+
+            <label for="gender">성별:</label>
+            <select name="gender" id="gender">
+                <option value="남성">남성</option>
+                <option value="여성">여성</option>
+                <option value="기타">기타</option>
+            </select><br />
+
+            <label for="birth">생년월일:</label>
+            <select name="year" id="birth-year">
+            </select>
+            <select name="month" id="birth-month">
+            </select>
+            <select name="day" id="birth-day">
+            </select><br />
+
+            <button type="button" onclick="userSignUp()">회원가입</button>
+        </form>
+    </div>
+    <br/>
+    <br/>
+    <button onclick="authKaKao()">카톡!</button> <br/>
+    <br/>
+</body>
+<script>
+    async function cookieAuthCheck(){
+        const res = await axios({
+            method:'post',
+            url:'/auth'
+        })
+        console.log("auth res",res);
+    }
+</script>
+<script src="/public/js/login.js" defer></script>
+</html>
