+++ views/profile.ejs
@@ -132,6 +132,7 @@
     </div>
 
     <script>
+        console.log('button',document.querySelector('.profile-edit-button'));
         // 프로필 수정 버튼 클릭 시 이벤트
         document.querySelector('.profile-edit-button').addEventListener('click', function () {
             // 프로필 수정 페이지로 이동 (URL을 수정 페이지의 경로로 변경)
